Updated README File
Updated the README file to reflect the current project's status.

I've also updated the file such that it includes a "Who is this for?" section instead of a "How well does this work?" section. It was a good idea to originally include it here, but I think it is more appropriate to be on the original Gifty repository exclusively.